+++ test/publish/test_publication.py
@@ -3,141 +3,7 @@ import pytest
 
 from rialto_airflow.publish import publication
 from rialto_airflow.database import Publication, Author, Funder
-from test.test_utils import TestRow
-
-
-def dim_json():
-    return {
-        "type": "article",
-        "doi": "10.000/000001",
-        "journal": {"title": "Delicious Limes Journal of Science"},
-        "issue": "12",
-        "pages": "1-10",
-        "volume": "1",
-        "mesh_terms": ["Delicions", "Limes"],
-        "pmid": "123",
-        "linkout": "https://example_dim.com",
-    }
-
-
-def dim_json_no_title():
-    return {"other_data": {"bogus": "no journal title here"}}
-
-
-def openalex_json():
-    return {
-        "type": "preprint",
-        "biblio": {"issue": "11", "first_page": "1", "last_page": "9", "volume": "2"},
-        "primary_location": {
-            "source": {
-                "type": "journal",
-                "display_name": "Ok Limes Journal of Science",
-            }
-        },
-        "locations": [
-            {"pdf_url": "https://example_openalex_pdf.com"},
-        ],
-        "mesh": [
-            {"descriptor_name": "Ok"},
-            {"descriptor_name": "Lemons"},
-        ],
-        "ids": {
-            "doi": "10.000/000001",
-            "pmid": "1234",
-        },
-    }
-
-
-def openalex_no_title_json():
-    return {
-        "locations": [
-            {
-                "bogus": {
-                    "nope": "not here",
-                }
-            },
-            {"source": {"type": "geography", "display_name": "New York"}},
-        ],
-    }
-
-
-def wos_json():
-    return {
-        "fullrecord_metadata": {
-            "normalized_doctypes": {"doctype": ["Article", "Abstract"]}
-        },
-        "static_data": {
-            "summary": {
-                "pub_info": {
-                    "pubyear": "2023",
-                    "issue": "10",
-                    "vol": "3",
-                    "page": {"begin": "1", "end": "8"},
-                },
-                "titles": {
-                    "count": 2,
-                    "title": [
-                        {"type": "source", "content": "Meh Limes Journal of Science"},
-                        {"type": "bogus", "content": "Bogus"},
-                    ],
-                },
-            }
-        },
-        "dynamic_data": {
-            "cluster_related": {
-                "identifiers": {
-                    "identifier": [
-                        {
-                            "type": "pmid",
-                            "value": "1234567",
-                        }
-                    ],
-                },
-            }
-        },
-    }
-
-
-def wos_json_no_page_info():
-    return {
-        "static_data": {
-            "summary": {
-                "pub_info": {
-                    "page": {"page_count": "7"},
-                },
-            }
-        }
-    }
-
-
-def sulpub_json():
-    return {
-        "journal": {
-            "name": "Bad Limes Journal of Science",
-            "issue": "9",
-            "pages": "1-7",
-            "volume": "4",
-        },
-        "pmid": "123456",
-    }
-
-
-def pubmed_json():
-    return {
-        "MedlineCitation": {
-            "Article": {
-                "ArticleTitle": "Example Title",
-            },
-        },
-        "PubmedData": {
-            "ArticleIdList": {
-                "ArticleId": [
-                    {"@IdType": "pubmed", "#text": "36857419"},
-                    {"@IdType": "doi", "#text": "10.1182/bloodadvances.2022008893"},
-                ]
-            },
-        },
-    }
+import test.publish.data as test_data
 
 
 @pytest.fixture
@@ -149,11 +15,11 @@ def dataset(test_session):
             apc=123,
             open_access="gold",
             pub_year=2023,
-            dim_json=dim_json(),
-            openalex_json=openalex_json(),
-            wos_json=wos_json(),
-            sulpub_json=sulpub_json(),
-            pubmed_json=pubmed_json(),
+            dim_json=test_data.dim_json(),
+            openalex_json=test_data.openalex_json(),
+            wos_json=test_data.wos_json(),
+            sulpub_json=test_data.sulpub_json(),
+            pubmed_json=test_data.pubmed_json(),
         )
 
         pub.authors.append(
@@ -463,284 +329,3 @@ def test_write_contributions_by_department(test_session, snapshot, dataset, capl
 
     assert "starting to write contributions by school/department" in caplog.text
     assert "finished writing contributions by school/department" in caplog.text
-
-
-def test_journal_with_dimensions_title():
-    row = TestRow(
-        dim_json=dim_json(),
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    assert publication._journal(row) == "Delicious Limes Journal of Science"
-
-
-def test_journal_missing_dimensions_title():
-    row = TestRow(
-        dim_json=dim_json_no_title(),
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from openalex
-    assert publication._journal(row) == "Ok Limes Journal of Science"
-
-
-def test_journal_without_dimensions_title():
-    row = TestRow(
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from openalex
-    assert publication._journal(row) == "Ok Limes Journal of Science"
-
-
-def test_journal_without_dimensions_and_missing_open_alex_title():
-    row = TestRow(
-        openalex_json=openalex_no_title_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from wos
-    assert publication._journal(row) == "Meh Limes Journal of Science"
-
-
-def test_journal_without_dimensions_or_open_alex_title():
-    row = TestRow(
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from wos
-    assert publication._journal(row) == "Meh Limes Journal of Science"
-
-
-def test_journal_without_dimensions_or_open_alex_or_wos_title():
-    row = TestRow(
-        sulpub_json=sulpub_json(),
-    )
-    # comes from sulpub
-    assert publication._journal(row) == "Bad Limes Journal of Science"
-
-
-def test_journal_with_dimensions_issue():
-    row = TestRow(
-        dim_json=dim_json(),
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    assert publication._issue(row) == "12"
-
-
-def test_journal_without_dimensions_issue():
-    row = TestRow(
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from openalex
-    assert publication._issue(row) == "11"
-
-
-def test_journal_without_dimensions_or_openalex_issue():
-    row = TestRow(
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from openalex
-    assert publication._issue(row) == "10"
-
-
-def test_journal_without_dimensions_or_openalex_or_wos_issue():
-    row = TestRow(
-        sulpub_json=sulpub_json(),
-    )
-    # comes from sulpub
-    assert publication._issue(row) == "9"
-
-
-def test_journal_with_dimensions_mesh():
-    row = TestRow(
-        dim_json=dim_json(),
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from dimensions
-    assert publication._mesh(row) == "Delicions|Limes"
-
-
-def test_journal_without_dimensions_mesh():
-    row = TestRow(
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from openalex
-    assert publication._mesh(row) == "Ok|Lemons"
-
-
-def test_journal_with_dimensions_pages():
-    row = TestRow(
-        dim_json=dim_json(),
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from dimensions
-    assert publication._pages(row) == "1-10"
-
-
-def test_journal_without_dimensions_pages():
-    row = TestRow(
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from openalex
-    assert publication._pages(row) == "1-9"
-
-
-def test_journal_without_dimensions_or_openalex_pages():
-    row = TestRow(
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from wos
-    assert publication._pages(row) == "1-8"
-
-
-def test_journal_missing_wos_page_info():
-    row = TestRow(
-        wos_json=wos_json_no_page_info(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from sulpub
-    assert publication._pages(row) == "1-7"
-
-
-def test_journal_without_dimensions_or_openalex_or_wos_pages():
-    row = TestRow(
-        sulpub_json=sulpub_json(),
-    )
-    # comes from sulpub
-    assert publication._pages(row) == "1-7"
-
-
-def test_journal_with_pubmed_pmid():
-    row = TestRow(
-        pubmed_json=pubmed_json(),
-        dim_json=dim_json(),
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from pubmed
-    assert publication._pmid(row) == "36857419"
-
-
-def test_journal_without_pubmed_pmid():
-    row = TestRow(
-        dim_json=dim_json(),
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from dimensions
-    assert publication._pmid(row) == "123"
-
-
-def test_journal_without_pubmed_or_dimensions_pmid():
-    row = TestRow(
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from openalex
-    assert publication._pmid(row) == "1234"
-
-
-def test_journal_without_pubmed_or_dimensions_or_openalex_pmid():
-    row = TestRow(
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from sulpub
-    assert publication._pmid(row) == "123456"
-
-
-def test_journal_without_pubmed_or_dimensions_or_openalex_or_sulpub_pmid():
-    row = TestRow(
-        wos_json=wos_json(),
-    )
-    # comes from wos
-    assert publication._pmid(row) == "1234567"
-
-
-def test_journal_with_dimensions_url():
-    row = TestRow(
-        dim_json=dim_json(),
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from dimensions
-    assert publication._url(row) == "https://example_dim.com"
-
-
-def test_journal_without_dimensions_url():
-    row = TestRow(
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from openalex
-    assert publication._url(row) == "https://example_openalex_pdf.com"
-
-
-def test_journal_without_dimensions_or_openalex_url():
-    row = TestRow(
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # no url
-    assert publication._url(row) is None
-
-
-def test_journal_with_dimensions_volume():
-    row = TestRow(
-        dim_json=dim_json(),
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from dimensions
-    assert publication._volume(row) == "1"
-
-
-def test_journal_without_dimensions_volume():
-    row = TestRow(
-        openalex_json=openalex_json(),
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from openalex
-    assert publication._volume(row) == "2"
-
-
-def test_journal_without_dimensions_or_openalex_volume():
-    row = TestRow(
-        wos_json=wos_json(),
-        sulpub_json=sulpub_json(),
-    )
-    # comes from wos
-    assert publication._volume(row) == "3"
-
-
-def test_journal_without_dimensions_or_openalex_or_wos_volume():
-    row = TestRow(
-        sulpub_json=sulpub_json(),
-    )
-    # comes from sulpub
-    assert publication._volume(row) == "4"
